# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paytm Clone API",
    "version": "1.0.0",
    "description": "A node.js Express API"
  },
  "paths": {
    "/api/v1/addbalance": {
      "post": {
        "summary": "Add balance to the wallet",
        "tags": [
          "Wallet"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "amount"
                ],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "ID of the user"
                  },
                  "amount": {
                    "type": "integer",
                    "description": "amount to be added to the wallet"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully added the balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "receiver": {
                          "type": "integer"
                        },
                        "sender": {
                          "type": "integer"
                        },
                        "amount": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Cannot communicate to the bank API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deductbalance": {
      "post": {
        "summary": "Deduct amount from the wallet",
        "tags": [
          "Wallet"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "amount"
                ],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "ID of the user"
                  },
                  "amount": {
                    "type": "integer",
                    "description": "amount to be deducted from the wallet"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully deducted the balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "sender": {
                          "type": "integer"
                        },
                        "bankIdReceiver": {
                          "type": "integer"
                        },
                        "amountTransferred": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Cannot communicate to the bank API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}